+++ Gemfile
@@ -44,7 +44,7 @@ gem "kamal", require: false
 gem "thruster", require: false
 
 # Using gem to detect browser type
-gem 'browser'
+gem "browser"
 
 # Use Active Storage variants [https://guides.rubyonrails.org/active_storage_overview.html#transforming-images]
 # gem "image_processing", "~> 1.2"
