+++ app/controllers/application_controller.rb
@@ -13,5 +13,4 @@ class ApplicationController < ActionController::Base
   def mobile_device?
     @browser.device.mobile?
   end
-
 end
