+++ app/views/kainga/index.html.erb
@@ -2,8 +2,10 @@
   <p class="text-left fs-4">Kam's Stuff</p>
   <figcaption class="blockquote-footer text-left">
     Various Coding Samples and Tools
-    <br/>Kia ora : Nau Mai Piki Mai
+    <br/>Kia ora : Nau Mai Piki Mai<br/>
+    <%= link_to "mobile Version", mobile_path %>
   </figcaption>
+
 </div>
 
 <div class="row">
