+++ app/views/kainga/mobile.html.erb
@@ -3,6 +3,7 @@
   <figcaption class="blockquote-footer text-start">
     Various Coding Samples and Tools
     <br/>Kia ora : Nau Mai Piki Mai
+    <br/><%= link_to "Kainga", root_path %> | <%= link_to "Home", root_path %>
   </figcaption>
 </div>
